fixes to activate and patch

+++ woven/main.py
@@ -3,7 +3,7 @@ import os,sys
 
 from django.utils.importlib import import_module
 
-from fabric.api import env, local, run, cd, sudo
+from fabric.api import env, local, run, cd, sudo, settings
 from fabric.main import find_fabfile
 from fabric.contrib.files import append, exists
 
@@ -48,7 +48,7 @@ def active_version():
     
     Just examine the which environment is symlinked
     """
-    link = '/'.join([env.deployment_root,env.project_name])
+    link = '/'.join([env.deployment_root,'env',env.project_name])
     if not exists(link): return None
     active = os.path.split(run('ls -al '+link).split(' -> ')[1])[1]
     return active
@@ -58,51 +58,55 @@ def activate():
     Activates the version or the current version if criteria is met
     """
     active = active_version()
-    stop_webservices()
-    if not active == env.project_version:
+
+    if env.patch or active <> env.project_fullname:
+        stop_webservices()
         
-        if not env.patch:
-            #TODO - DATA MIGRATION HERE
-    
-            #delete existing symlink
-            run('rm -f '+os.path.join(env.deployment_root,'env',env.project_name))
-            run('ln -s %s %s'% (os.path.join(env.deployment_root,'env',env.project_fullname),
-                                os.path.join(env.deployment_root,'env',env.project_name))
-               )
-            #create shortcuts for virtualenv activation in the home directory
-            activate_env = '/'.join(['/home',env.user,'workon-'+env.project_name])
-            if not exists(activate_env):
-                run("touch "+activate_env)
-                append('#!/bin/bash',activate_env)
-                append("source "+ os.path.join(env.deployment_root,'env',env.project_name,'bin','activate'),
-                       activate_env)
-                append("cd "+ os.path.join(env.deployment_root,'env',env.project_name,'project',env.project_name),
-                       activate_env)
-                run("chmod +x "+activate_env)
-            
-            #activate sites
-            #enabled_sites = _ls_sites('/etc/apache2/sites-enabled') + _ls_sites('/etc/nginx/sites-enabled')
-            activate_sites = [''.join([d.replace('.','_'),'-',env.project_version,'.conf']) for d in env.DOMAINS]
-            site_paths = ['/etc/apache2','/etc/nginx']
-            
-            #disable existing sites
-            for path in site_paths:
-                for site in _ls_sites('/'.join([path,'sites-enabled'])):
-                    if site not in activate_sites:
-                        sudo("rm %s/sites-enabled/%s"% (path,site))
-            
-            #activate new sites
-            for path in site_paths:
-                for site in activate_sites:
-                    if not exists('/'.join([path,'sites-enabled',site])):
-                        sudo("chmod 644 %s" % '/'.join([path,'sites-available',site]))
-                        sudo("ln -s %s/sites-available/%s %s/sites-enabled/%s"% (path,site,path,site))
-      
+    if not env.patch and active <> env.project_fullname:
+        #TODO - DATA MIGRATION HERE
+
+        #delete existing symlink
+        run('rm -f '+os.path.join(env.deployment_root,'env',env.project_name))
+        run('ln -s %s %s'% (os.path.join(env.deployment_root,'env',env.project_fullname),
+                            os.path.join(env.deployment_root,'env',env.project_name))
+           )
+        #create shortcuts for virtualenv activation in the home directory
+        activate_env = '/'.join(['/home',env.user,'workon-'+env.project_name])
+        if not exists(activate_env):
+            run("touch "+activate_env)
+            append('#!/bin/bash',activate_env)
+            append("source "+ os.path.join(env.deployment_root,'env',env.project_name,'bin','activate'),
+                   activate_env)
+            append("cd "+ os.path.join(env.deployment_root,'env',env.project_name,'project',env.project_name),
+                   activate_env)
+            run("chmod +x "+activate_env)
         
-        print env.host,env.project_fullname, "ACTIVATED"
+        #activate sites
+        #enabled_sites = _ls_sites('/etc/apache2/sites-enabled') + _ls_sites('/etc/nginx/sites-enabled')
+        activate_sites = [''.join([d.replace('.','_'),'-',env.project_version,'.conf']) for d in env.DOMAINS]
+        site_paths = ['/etc/apache2','/etc/nginx']
+        
+        #disable existing sites
+        for path in site_paths:
+            for site in _ls_sites('/'.join([path,'sites-enabled'])):
+                if site not in activate_sites:
+                    sudo("rm %s/sites-enabled/%s"% (path,site))
+        
+        #activate new sites
+        for path in site_paths:
+            for site in activate_sites:
+                if not exists('/'.join([path,'sites-enabled',site])):
+                    sudo("chmod 644 %s" % '/'.join([path,'sites-available',site]))
+                    sudo("ln -s %s/sites-available/%s %s/sites-enabled/%s"% (path,site,path,site))
+  
+        if env.verbosity:
+            print env.host,env.project_fullname, "ACTIVATED"
     else:
-        print env.project_fullname,"is the active version"
-    start_webservices()
+        if env.verbosity and not env.patch:
+            print env.project_fullname,"is the active version"
+    if env.patch or active <> env.project_fullname:
+        start_webservices()
+        print
     
 
 def deploy():
@@ -119,8 +123,9 @@ def patch():
     """
     Patch the current version. Does not install packages or delete files
     """
-    with setting(patch=True):
+    with settings(patch=True):
         deploy()
+        activate()
     
 
 def setup_environ(settings=None, setup_dir=''):
